修复global_ripple.dart startTime类型错误(int->double)

+++ chumian_app/lib/widgets/global_ripple.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 
 class _RippleData {
   final Offset position;
-  final double startTime;
+  final int startTime;
   _RippleData({required this.position, required this.startTime});
 }
